Move distributed_util.py to test/
